requirements.txt
# Test Data Generator Requirements
mysql-connector-python==8.2.0
